requirements.txt
numpy==1.19.2
opencv-contrib-python==4.3.0.36
opencv-python==4.1.1.26
pandas==0.25.3
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
